shader_model should have reflection bump map defined
These fields are unused by the game, itself, but the tools do interact
with these fields on build. Therefore, they should be documented and
put in the tag, even if they are effectively just cache only.

These are apparently used in Stubbs the Zombie, although we do not
presently support this game.